Better condition order updating during recalculate

+++ core/app/services/spree/cart/recalculate.rb
@@ -4,12 +4,14 @@ module Spree
       prepend Spree::ServiceModule::Base
 
       def call(order:, line_item:, line_item_created: false, options: {})
-        order.remove_gift_card if order.gift_card.present?
         order_updater = ::Spree::OrderUpdater.new(order)
 
-        order.payments.store_credits.checkout.destroy_all
-        order_updater.update_totals
-        order_updater.persist_totals
+        if order.gift_card.present? || order.payments.store_credits.checkout.any?
+          order.remove_gift_card if order.gift_card.present?
+          order.payments.store_credits.checkout.destroy_all if order.payments.store_credits.checkout.any?
+          order_updater.update_totals
+          order_updater.persist_totals
+        end
 
         shipment = options[:shipment]
         if shipment.present?
